Adjust to different Regex diagnostic in different Java versions
